feat: UserAvatar to UserGroup 파일 병합
